add precode_test file and run test
