+++ precode_test.go
@@ -1,0 +1,51 @@
+package main
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
+	totalCount := 4
+
+	req := httptest.NewRequest("GET", "/cafe?count=10&city=moscow", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	assert.Equal(t, http.StatusOK, responseRecorder.Code)
+	responseBody := responseRecorder.Body.String()
+	sliceCafe := strings.Split(responseBody, ",")
+	require.Len(t, sliceCafe, totalCount)
+
+}
+
+func TestMainHandlerWhenWrongCity(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?count=7&city=piter", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	assert.Equal(t, http.StatusBadRequest, responseRecorder.Code)
+	require.NotEmpty(t, responseRecorder.Body)
+	require.Equal(t, responseRecorder.Body.String(), "wrong city value")
+}
+
+func TestMainHandlerWhenOk(t *testing.T) {
+
+	req := httptest.NewRequest("GET", "/cafe?count=7&city=moscow", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	assert.Equal(t, http.StatusOK, responseRecorder.Code)
+	require.NotEmpty(t, responseRecorder.Body)
+}
